+++ observers/tests/FloorContactTest.cpp
@@ -30,12 +30,28 @@ using palimpsest::KeyError;
 class FloorContactTest : public ::testing::Test {
  protected:
   void SetUp() override {
+    config_("floor_contact")("upper_leg_torque_threshold") = 10.0;
+    config_("wheel_contact")("liftoff_inertia") = 0.001;
+    config_("wheel_contact")("min_touchdown_acceleration") = 2.0;
+    config_("wheel_contact")("min_touchdown_torque") = 0.015;
+    config_("wheel_contact")("touchdown_inertia") = 0.004;
+
+    // We set the cutoff period close to dt_ to generate accelerations larger
+    // than the min touchdown acceleration more easily in some tests.
+    config_("wheel_contact")("cutoff_period") = 3 * dt_;
+
     FloorContact::Parameters params;
+    params.dt = dt_;
+    params.upper_leg_joints = {"hip", "knee"};
+    params.wheels = {"big_wheel", "small_wheel"};
+
     floor_contact_ = std::make_unique<FloorContact>(params);
   }
 
  protected:
-  //! Test config
+  const double dt_ = 1.0 / 250.0;
+
+  //! Configuration
   Dictionary config_;
 
   //! Observer
@@ -43,11 +59,60 @@ class FloorContactTest : public ::testing::Test {
 };
 
 TEST_F(FloorContactTest, WriteEvenAfterEmptyRead) {
-  // Writing after an empty read is the current spec though not ideal
+  // Writing after an empty read is the current spec, though not ideal
   Dictionary observation;
   ASSERT_NO_THROW(floor_contact_->read(observation));
   ASSERT_NO_THROW(floor_contact_->write(observation));
   ASSERT_FALSE(observation.is_empty());
+}
+
+TEST_F(FloorContactTest, UnconfiguredDoesNotThrow) {
+  FloorContact::Parameters params;
+  FloorContact observer(params);
+  Dictionary observation;
+  ASSERT_NO_THROW(observer.read(observation));
+}
+
+TEST_F(FloorContactTest, PartialObservationDoesNotThrow) {
+  Dictionary observation;
+  observation("servo")("big_wheel")("velocity") = 0.0;
+  ASSERT_NO_THROW(floor_contact_->read(observation));
+}
+
+TEST_F(FloorContactTest, NoTorqueNoContact) {
+  floor_contact_->reset(config_);
+
+  Dictionary observation;
+  observation("servo")("big_wheel")("velocity") = 0.1;
+  observation("servo")("big_wheel")("torque") = 0.0;
+  observation("servo")("small_wheel")("velocity") = 0.1;
+  observation("servo")("small_wheel")("torque") = 0.0;
+
+  ASSERT_NO_THROW(floor_contact_->read(observation));
+  ASSERT_NO_THROW(floor_contact_->write(observation));
+  ASSERT_FALSE(observation("floor_contact").get<bool>("contact"));
+}
+
+TEST_F(FloorContactTest, BigTorqueMeansContact) {
+  floor_contact_->reset(config_);
+
+  Dictionary observation;
+
+  const double vel = 10.0;  // rad/s
+  observation("servo")("big_wheel")("velocity") = vel;
+  observation("servo")("big_wheel")("torque") = 10.0;
+  observation("servo")("small_wheel")("velocity") = vel;
+  observation("servo")("small_wheel")("torque") = 10.0;
+  floor_contact_->read(observation);
+
+  // Increase velocity to produce accel > min_touchdown_acceleration
+  const double new_vel = vel + 5.0 * dt_;
+  observation("servo")("big_wheel")("velocity") = new_vel;
+  observation("servo")("small_wheel")("velocity") = new_vel;
+  floor_contact_->read(observation);
+
+  ASSERT_NO_THROW(floor_contact_->write(observation));
+  ASSERT_TRUE(observation("floor_contact").get<bool>("contact"));
 }
 
 }  // namespace upkie::observers::tests